Minor improvements and refactoring on account-details
