+++ account-details/src/main/java/ba/klika/focus/reactive/accounts/accountdetails/accountdetails/service/IAccountDetailsService.java
@@ -1,13 +1,11 @@
-package ba.klika.focus.reactive.accounts.accountdetails.accountdetails;
+package ba.klika.focus.reactive.accounts.accountdetails.accountdetails.service;
 
 import ba.klika.focus.reactive.accounts.accountdetails.accountdetails.database.Account;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
 public interface IAccountDetailsService {
-
-    Flux<Account> getAllAccount();
+    Flux<Account> getAllAccounts();
 
     Mono<Account> getAccountByIban(String iban);
-
 }
